Enhaced the Ui interface for both admin and author

+++ src/main/java/com/smartblog/ui/view/performance/PerformanceViewModel.java
@@ -99,7 +99,7 @@ public class PerformanceViewModel {
         return benchmark.size();
     }
 
-    public long getTotalDuration() {
+    public double getTotalDuration() {
         return benchmark.generateReport().getTotalDuration();
     }
